現在のタスク部分をコンポーネント化
